Add various algorithms and data structures implementations
- Implement Kahn's algorithm for topological sorting in topological_kahn.cpp
- Create a tree graph representation and calculate child counts and heights in tree_graph.cpp
- Add binary exponentiation algorithm in binary_pow.cpp
- Implement string Fibonacci sequence generator in string_fibonacci.cpp
- Solve the coin change problem using dynamic programming in coin_problem.cpp
- Add O(n^2) Longest Increasing Subsequence (LIS) algorithm in lis_n2.cpp
- Update CMakeLists.txt to correct directory name for data structures
- Remove outdated difference array, prefix sum, and sorting implementations
- Implement various string and vector operations in respective files

+++ leetcode/backtracking/46_permutations.cpp
@@ -1,3 +1,4 @@
+#include <cstddef>
 #include <iostream>
 using namespace std;
 
@@ -60,8 +61,8 @@ int main() {
             cin >> nums[i];
 
         vector<vector<int>> perms = permute(nums);
-        for (int i = 0; i < perms.size(); i++) {
-            for (int j = 0; j < perms[i].size(); j++)
+        for (size_t i = 0; i < perms.size(); i++) {
+            for (size_t j = 0; j < perms[i].size(); j++)
                 cout << perms[i][j] << " ";
             cout << endl;
         }
